fix: address CodeRabbit issues and resolve all clippy warnings
- Fix relay success_rate display using correct 0-100 thresholds
- Make protected event rejection explicit for NIP-70 compliance
- Fix duplicated cfg attribute in indexeddb_database.rs
- Replace Arc with Rc for single-threaded WASM (IdbDatabase)
- Use const initializers for thread_local values
- Remove unnecessary clone() on Copy types (Signals)
- Collapse else-if blocks and fix match_result_ok patterns
- Replace useless vec! macros with arrays
- Add type alias to reduce type complexity in use_infinite_scroll

+++ src/routes/home.rs
@@ -432,7 +432,7 @@ pub fn Home(list: String) -> Element {
                     #[cfg(target_arch = "wasm32")]
                     {
                         use gloo_timers::future::TimeoutFuture;
-                        TimeoutFuture::new((batch_idx as u32 * BATCH_DELAY_MS as u32) as u32).await;
+                        TimeoutFuture::new(batch_idx as u32 * BATCH_DELAY_MS as u32).await;
                     }
 
                     #[cfg(not(target_arch = "wasm32"))]
